move config to field_friend module
