+++ field_friend/__init__.py
@@ -1,8 +1,9 @@
-from . import interface, log_configuration
+from . import config, interface, log_configuration
 from .system import System
 
 __all__ = [
     'System',
+    'config',
     'interface',
     'log_configuration',
 ]
